test structure refactor
